Cast output data to uint8 by default (#299)
* update ImageData.post_process to cast ouptut data to uint8 when doing rescaling

* update test

+++ tests/test_io_cogeo.py
@@ -387,7 +387,7 @@ def test_imageData_output():
 
         imgr = img.post_process(in_range=(img.data.min(), img.data.max()))
         assert not numpy.array_equal(img.data, imgr.data)
-        assert imgr.data.dtype == "uint16"
+        assert imgr.data.dtype == "uint8"
         assert imgr.data.min() == 0
         assert imgr.data.max() == 255
         assert imgr.bounds == img.bounds
